requirements.txt
numpy>=1.24
pandas>=2.0
scipy>=1.10
statsmodels>=0.14

matplotlib>=3.7
seaborn>=0.13
missingno>=0.5

scikit-learn>=1.3

xgboost>=2.0
catboost>=1.2
lightgbm>=4.0
ipykernel>=6.24
streamlit>=1.42
pyyaml
mlflow
pytest
dvc